fix: 修复服务器端TypeScript严格模式下的类型兼容性问题
- 在AudioVisualizer.tsx中为getByteFrequencyData添加类型断言
- 在useQwenSpeechRecognition.ts中为resampleAudio参数添加类型断言
- 解决ArrayBufferLike与ArrayBuffer的类型不匹配问题
- 确保代码在不同TypeScript版本下都能正常编译

+++ frontend/src/components/AudioVisualizer.tsx
@@ -100,7 +100,7 @@ const AudioVisualizer: React.FC<AudioVisualizerProps> = ({
     if (!ctx) return
 
     // 获取音频数据
-    analyserRef.current.getByteFrequencyData(dataArrayRef.current)
+    analyserRef.current.getByteFrequencyData(dataArrayRef.current as Uint8Array)
 
     // 清空画布
     ctx.clearRect(0, 0, width, height)
